cloud-functions/document-ai-processor/requirements.txt
functions-framework==3.5.0
google-cloud-documentai==2.20.0
google-cloud-storage==2.10.0
pyyaml==6.0.1